# docs/swagger.yaml
basePath: /api/v1
...
host: localhost:3000
info:
  contact:
    email: support@example.com
    name: API Support
  description: PBI API Documentation
  termsOfService: http://swagger.io/terms/
  title: PBI API
  version: "1.0"
paths:
...
  /product:
    get:
      description: Get product list with filters & pagination
      parameters:
      - description: Product name
        in: query
        name: nama_produk
        type: string
      - description: Category ID
        in: query
        name: category_id
        type: integer
      - description: Toko ID
        in: query
        name: toko_id
        type: integer
      - description: Minimum price
        in: query
        name: min_harga
        type: number
      - description: Maximum price
        in: query
        name: max_harga
        type: number
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Limit per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success get product list
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      summary: Get all products
      tags:
      - Product
    post:
      consumes:
      - multipart/form-data
      description: Create new product with images (multipart/form-data)
      parameters:
      - description: Product name
        in: formData
        name: nama_produk
        required: true
        type: string
      - description: Product slug
        in: formData
        name: slug
        required: true
        type: string
      - description: Product description
        in: formData
        name: deskripsi
        required: true
        type: string
      - description: Category ID
        in: formData
        name: id_category
        required: true
        type: integer
      - description: Reseller price
        in: formData
        name: harga_reseller
        required: true
        type: number
      - description: Consumer price
        in: formData
        name: harga_konsumen
        required: true
        type: number
      - description: Stock
        in: formData
        name: stok
        required: true
        type: integer
      - description: Product images (multiple)
        in: formData
        name: photos
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Success create product
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Create product
      tags:
      - Product
  /product/{id}:
    delete:
      description: Delete product by ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success delete product
          schema:
            type: object
        "400":
          description: Invalid product ID
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "404":
          description: Product not found
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Delete product
      tags:
      - Product
    get:
      description: Get product detail
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success get product detail
          schema:
            type: object
        "400":
          description: Invalid product ID
          schema:
            type: object
        "404":
          description: Product not found
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      summary: Get product by ID
      tags:
      - Product
    put:
      consumes:
      - multipart/form-data
      description: Update product by ID (multipart/form-data)
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Product name
        in: formData
        name: nama_produk
        type: string
      - description: Product slug
        in: formData
        name: slug
        type: string
      - description: Product description
        in: formData
        name: deskripsi
        type: string
      - description: Category ID
        in: formData
        name: id_category
        type: integer
      - description: Reseller price
        in: formData
        name: harga_reseller
        type: number
      - description: Consumer price
        in: formData
        name: harga_konsumen
        type: number
      - description: Stock
        in: formData
        name: stok
        type: integer
      - description: Product images (multiple)
        in: formData
        name: photos
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Success update product
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "404":
          description: Product not found
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Update product
      tags:
      - Product
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"